requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
autocommand==2.2.2
backports.csv==1.0.7
beautifulsoup4==4.13.5
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cheroot==10.0.1
CherryPy==18.10.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.16
decorator==5.2.1
docx==0.2.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
feedparser==6.0.11
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.collections==5.2.1
jaraco.context==6.0.1
jaraco.functools==4.3.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.3
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==6.0.1
marisa-trie==1.2.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==10.7.0
murmurhash==1.0.13
mysqlclient==2.2.7
namex==0.1.0
nest-asyncio==1.6.0
nltk==3.9.1
node==1.2.2
npm==0.1.1
numpy==2.3.0
odict==1.9.0
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.1
pattern3==3.0.0
pdfminer.six==20250506
pdfminer3k==1.3.4
pillow==11.2.1
platformdirs==4.4.0
plumber==1.7
ply==3.11
portend==3.2.1
preshed==3.0.10
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
pyzmq==27.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.5
tempora==5.8.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.5
zc.lockfile==3.0.post1
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.1
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
